# contracts/openapi/ai-study-mentor.v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EL English Mentor API",
    "version": "1.0.0",
    "description": "Canonical cross-client contract for authentication, profile, shared settings, the learning catalog, immutable learning events, and server-derived progress."
  },
  "servers": [
    {
      "url": "https://api.example.com/api/v1",
      "description": "Production placeholder; deployment configuration supplies the actual hostname."
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "register",
        "summary": "Register a user and create a session",
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientPlatform"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/AuthSessionCreated"
          },
          "409": {
            "$ref": "#/components/responses/EmailAlreadyRegistered"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "login",
        "summary": "Authenticate and create a session",
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientPlatform"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthSession"
          },
          "401": {
            "$ref": "#/components/responses/InvalidCredentials"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "refreshSession",
        "summary": "Rotate a refresh session and return a new access token",
        "description": "For `web`, the request body MUST be omitted and the refresh token is supplied only by a Secure, HttpOnly, SameSite cookie. For `android`, the request body is required and contains the refresh token obtained from secure device storage. A successful Web response never includes a refresh token in JSON; the rotated token is set only in Set-Cookie. A successful Android response includes the rotated refresh token in its JSON body for storage in a Keystore-backed encrypted store.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientPlatform"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AndroidRefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/RefreshedSession"
          },
          "401": {
            "$ref": "#/components/responses/RefreshRejected"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "logoutCurrentSession",
        "summary": "Revoke the current refresh-token family",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientPlatform"
          }
        ],
        "responses": {
          "204": {
            "description": "Current session revoked. Web clients also receive a clearing Set-Cookie header."
          },
          "401": {
            "$ref": "#/components/responses/SessionExpired"
          }
        }
      }
    },
    "/auth/logout-all": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "logoutAllSessions",
        "summary": "Revoke all active refresh-token families for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientPlatform"
          }
        ],
        "responses": {
          "204": {
            "description": "All sessions revoked."
          },
          "401": {
            "$ref": "#/components/responses/SessionExpired"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "operationId": "getCurrentUser",
        "summary": "Get the authenticated user identity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/CurrentUser"
          },
          "401": {
            "$ref": "#/components/responses/SessionExpired"
          }
        }
      }
    },
    "/me/profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "operationId": "getMyProfile",
        "summary": "Get the authenticated user's profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Profile"
          },
          "401": {
            "$ref": "#/components/responses/SessionExpired"
          }
        }
      },
      "patch": {
        "tags": [
          "Profile"
        ],
        "operationId": "updateMyProfile",
        "summary": "Update the authenticated user's profile",
        "description": "The caller must send the profile revision returned by GET as If-Match. A 409 response means the profile changed on another client and must be re-read before retrying.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Profile"
          },
          "401": {
            "$ref": "#/components/responses/SessionExpired"
          },
          "409": {
            "$ref": "#/components/responses/RevisionConflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/me/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "operationId": "getMySettings",
        "summary": "Get minimal synchronized learning settings",
        "description": "Theme, sound, reduced-motion, local push permission, and other device presentation state are intentionally not part of this initial synchronized contract.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Settings"
          },
          "401": {
            "$ref": "#/components/responses/SessionExpired"
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "operationId": "replaceMySettings",
        "summary": "Replace minimal synchronized learning settings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Settings"
          },
          "401": {
            "$ref": "#/components/responses/SessionExpired"
          },
          "409": {
            "$ref": "#/components/responses/RevisionConflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/me/progress": {
      "get": {
        "tags": [
          "Progress"
        ],
        "operationId": "getMyProgress",
        "summary": "Get the minimal server-authoritative progress read model",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ProgressSummary"
          },
          "401": {
            "$ref": "#/components/responses/SessionExpired"
          }
        }
      }
    },
    "/me/engagement": {
      "get": {
        "tags": [
          "Engagement"
        ],
        "operationId": "getMyEngagement",
        "summary": "Get backend-derived XP, level, streak, private achievements, and missions",
        "description": "Read-only projection. Clients must not calculate, persist, or submit engagement state.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Engagement"
          },
          "401": {
            "$ref": "#/components/responses/SessionExpired"
          }
        }
      }
    },
    "/me/campaign": {
      "get": {
        "tags": ["Engagement"],
        "operationId": "getMyCampaign",
        "summary": "Get the private backend-derived learning campaign map",
        "description": "Read-only deterministic map over active catalog records and accepted lesson completions. The catalog remains fully open; recommendedNodeId is guidance only and is not a lock, reward, or mastery claim.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/Campaign" },
          "401": { "$ref": "#/components/responses/SessionExpired" }
        }
      }
    },
...
    "/subjects": {
      "get": {
        "tags": [
          "Learning catalog"
        ],
        "operationId": "listSubjects",
        "summary": "List active subjects in display order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Subjects"
          },
          "401": {
            "$ref": "#/components/responses/SessionExpired"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived access JWT supplied in the Authorization header."
      }
    },
    "parameters": {
      "ClientPlatform": {
        "name": "X-Client-Platform",
        "in": "header",
        "required": true,
        "description": "Determines refresh-token transport only: web uses a Secure HttpOnly cookie; android uses the documented JSON field and secure device storage.",
        "schema": {
          "type": "string",
          "enum": [
            "web",
            "android"
          ]
        }
      },
...
    },
...
    "responses": {
      "AuthSessionCreated": {
        "description": "Registered and authenticated.",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          },
          "Set-Cookie": {
            "$ref": "#/components/headers/RefreshCookie"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SessionEnvelope"
            }
          }
        }
      },
...
      "EmailAlreadyRegistered": {
        "$ref": "#/components/responses/AuthError"
      },
...
    },
...
  }
}